Create MailChimp Audience Scraper
For the life of me, I was not able to download a CSV of an audience I needed to access in MailChimp. 
I created a web scraper in Python paired with MailChimp's API to solve this problem.
Specifics are omitted for data privacy, but the basics should be understood pretty easily.
Enjoy!